src/pyproject.toml
[tool.poetry]
name = "trucktrack"
version = "0.1.0"
description = ""
authors = ["Sherif Ragimov <ragimov700@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.11"
flake8 = "^7.0.0"
isort = "^5.13.2"
djangorestframework = "^3.15.1"
geopy = "^2.4.1"
factory-boy = "^3.3.0"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
redis = "^5.0.3"
django-celery-beat = "^2.6.0"
django-debug-toolbar = "^4.3.0"
drf-spectacular = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
